+++ experimental/css-has-pseudo/src/browser.js
@@ -21,7 +21,7 @@ export default function cssHasPseudo(document, options) {
 		if (!options.forcePolyfill) {
 			try {
 				// Chrome does not support forgiving selector lists in :has()
-				global.document.querySelector(':has(*, :does-not-exist, > *)');
+				document.querySelector(':has(*, :does-not-exist, > *)');
 
 				// Safari incorrectly returns the html element with this query
 				if (!document.querySelector(':has(:scope *)')) {