+++ plugins/css-prefers-color-scheme/.tape.mjs
@@ -5,18 +5,6 @@ postcssTape(plugin)({
 	'basic': {
 		message: 'supports basic usage'
 	},
-	'basic:color': {
-		message: 'supports { mediaQuery: "color" } usage',
-		options: {
-			mediaQuery: 'color'
-		}
-	},
-	'basic:color-index': {
-		message: 'supports { mediaQuery: "color-index" } usage',
-		options: {
-			mediaQuery: 'color-index'
-		}
-	},
 	'basic:preserve': {
 		message: 'supports { preserve: true } usage',
 		options: {
@@ -28,5 +16,11 @@ postcssTape(plugin)({
 		options: {
 			preserve: false
 		}
-	}
+	},
+	'browser': {
+		message: 'css for browser tests',
+		options: {
+			preserve: false
+		}
+	},
 });