+++ plugins/css-prefers-color-scheme/src/browser.js
@@ -1,5 +1,4 @@
 /* global document,window,matchMedia */
-const colorIndexRegExp = /((?:not )?all and )?(\(color-index: *(22|48|70)\))/i;
 const prefersColorSchemeRegExp = /prefers-color-scheme:/i;
 
 const prefersColorSchemeInit = initialColorScheme => {
@@ -24,50 +23,38 @@ const prefersColorSchemeInit = initialColorScheme => {
 		}
 
 		[].forEach.call(document.styleSheets || [], styleSheet => {
+			try {
+				// cssRules is a live list. Converting to an Array first.
+				const rules = [];
+				[].forEach.call(styleSheet.cssRules || [], cssRule => {
+					rules.push(cssRule);
+				});
 
-			// cssRules is a live list. Converting to an Array first.
-			const rules = [];
-			[].forEach.call(styleSheet.cssRules || [], cssRule => {
-				rules.push(cssRule);
-			});
+				rules.forEach(cssRule => {
+					const colorSchemeMatch = prefersColorSchemeRegExp.test(Object(cssRule.media).mediaText);
 
-			rules.forEach(cssRule => {
-				const colorSchemeMatch = prefersColorSchemeRegExp.test(Object(cssRule.media).mediaText);
-
-				if (colorSchemeMatch) {
-					const index = [].indexOf.call(cssRule.parentStyleSheet.cssRules, cssRule);
-					cssRule.parentStyleSheet.deleteRule(index);
-				} else {
-					const colorIndexMatch = (Object(cssRule.media).mediaText || '').match(colorIndexRegExp);
-
-					if (colorIndexMatch) {
-						// Old style which has poor browser support and can't handle complex media queries.
-						cssRule.media.mediaText = (
-							(/^dark$/i.test(colorScheme)
-								? colorIndexMatch[3] === '48'
-								: /^light$/i.test(colorScheme)
-									? colorIndexMatch[3] === '70'
-									: colorIndexMatch[3] === '22')
-								? 'not all and '
-								: ''
-						) + cssRule.media.mediaText.replace(colorIndexRegExp, '$2');
+					if (colorSchemeMatch) {
+						const index = [].indexOf.call(cssRule.parentStyleSheet.cssRules, cssRule);
+						cssRule.parentStyleSheet.deleteRule(index);
 					} else {
 						// New style which supports complex media queries.
-						const colorDepthMatch = (Object(cssRule.media).mediaText || '').match(/\( *(?:color|max-color): *(48842621|70318723|22511989) *\)/i);
+						const colorDepthMatch = (Object(cssRule.media).mediaText || '').match(/\( *(?:color|max-color): *(48842621|70318723) *\)/i);
 						if (colorDepthMatch && colorDepthMatch.length > 1) {
-							if (/^dark$/i.test(colorScheme) && (colorDepthMatch[1] === '48842621' || colorDepthMatch[1] === '22511989')) {
-								// No preference or preferred is dark and rule is dark.
-								cssRule.media.mediaText = cssRule.media.mediaText.replace(/\( *color: *(?:48842621|70318723) *\)/i, `(max-color: ${colorDepthMatch[1]})`);
-							} else if (/^light$/i.test(colorScheme) && (colorDepthMatch[1] === '70318723' || colorDepthMatch[1] === '22511989')) {
-								// No preference or preferred is light and rule is light.
-								cssRule.media.mediaText = cssRule.media.mediaText.replace(/\( *color: *(?:48842621|22511989) *\)/i, `(max-color: ${colorDepthMatch[1]})`);
+							if (colorScheme === 'dark' && (colorDepthMatch[1] === '48842621')) {
+								// preferred is dark and rule is dark.
+								cssRule.media.mediaText = cssRule.media.mediaText.replace(/\( *color: *(?:48842621) *\)/i, `(max-color: ${colorDepthMatch[1]})`);
+							} else if (colorScheme === 'light' && (colorDepthMatch[1] === '70318723')) {
+								// preferred is light and rule is light.
+								cssRule.media.mediaText = cssRule.media.mediaText.replace(/\( *color: *(?:70318723) *\)/i, `(max-color: ${colorDepthMatch[1]})`);
 							} else {
-								cssRule.media.mediaText = cssRule.media.mediaText.replace(/\( *max-color: *(?:48842621|70318723|22511989) *\)/i, `(color: ${colorDepthMatch[1]})`);
+								cssRule.media.mediaText = cssRule.media.mediaText.replace(/\( *max-color: *(?:48842621|70318723) *\)/i, `(color: ${colorDepthMatch[1]})`);
 							}
 						}
 					}
-				}
-			});
+				});
+			} catch (_) {
+				// Do nothing.
+			}
 		});
 	};
 	const result = Object.defineProperty(